requirements.txt
Flask==3.0.2
requests==2.31.0
scikit-learn==1.4.1
numpy==1.26.4
pandas==2.2.0
plotly==5.18.0
lxml==5.1.0
python-dotenv==1.0.1